Split of ISC_Admin
Split `ISC_Admin` in multiple classes and traits for more separation of module code.
