+++ admin/templates/header.php
@@ -13,7 +13,7 @@
 	</div>
 	<div id="isc-header-links">
 		<?php if ( ! \ISC\Plugin::is_pro() ) :
-			echo ISC_Admin::get_pro_link( 'header-pro' );
+			echo ISC\Admin_Utils::get_pro_link( 'header-pro' );
 		endif; ?>
 		<?php switch ( $screen_id ) :
 			case 'settings_page_isc-settings' : ?>
@@ -26,7 +26,7 @@
 				<a href="<?php echo esc_url( admin_url( 'upload.php?page=isc-sources' ) ); ?>"><?php esc_html_e( 'Tools', 'image-source-control-isc' ); ?></a>
 				<a href="<?php echo esc_url( admin_url( 'options-general.php?page=isc-settings' ) ); ?>"><?php esc_html_e( 'Settings', 'image-source-control-isc' ); ?></a>
 		<?php endswitch; ?>
-		<a href="<?php echo esc_url( ISC_Admin::get_manual_url( 'header-manual' ) ); ?>" target="_blank"><?php esc_html_e( 'Manual', 'image-source-control-isc' ); ?></a>
+		<a href="<?php echo esc_url( ISC\Admin_Utils::get_manual_url( 'header-manual' ) ); ?>" target="_blank"><?php esc_html_e( 'Manual', 'image-source-control-isc' ); ?></a>
 	</div>
 </div>
 <div class="wrap">
