+++ includes/isc.class.php
@@ -109,37 +109,9 @@ class ISC_Class {
 		 * @return string[]
 		 */
 		public function default_options() {
-			_deprecated_function( __METHOD__, '3.0.0', 'ISC\Plugin::default_options' );
+			_deprecated_function( __METHOD__, '3.0.0', 'ISC\Options::default_options' );
 
-			include ISCPATH . 'includes/default-licenses.php';
-			$default['display_type']              = [ 'list' ];
-			$default['list_on_archives']          = false;
-			$default['list_on_excerpts']          = false;
-			$default['image_list_headline']       = __( 'image sources', 'image-source-control-isc' );
-			$default['version']                   = ISCVERSION;
-			$default['images_per_page']           = 99999;
-			$default['thumbnail_in_list']         = false;
-			$default['thumbnail_size']            = 'thumbnail';
-			$default['thumbnail_width']           = 150;
-			$default['thumbnail_height']          = 150;
-			$default['warning_onesource_missing'] = true;
-			$default['remove_on_uninstall']       = false;
-			$default['caption_position']          = 'top-left';
-			$default['caption_style']             = null;
-			$default['source_pretext']            = __( 'Source:', 'image-source-control-isc' );
-			$default['enable_licences']           = false;
-			$default['licences']                  = apply_filters( 'isc-licences-list', $isc_default_licenses );
-			$default['list_included_images']      = '';
-			$default['overlay_included_images']   = '';
-			$default['block_options']             = true;
-			$default['enable_log']                = false;
-			$default['standard_source']           = 'custom_text';
-			$default['standard_source_text']      = '';
-
-			/**
-			 * Allow manipulating defaults for plugin settings
-			 */
-			return apply_filters( 'isc_default_settings', $default );
+			return ISC\Options::default_options();
 		}
 
 		/**
@@ -148,7 +120,7 @@ class ISC_Class {
 		 * @return string[]
 		 */
 		public function get_isc_options() {
-			_deprecated_function( __METHOD__, '3.0.0', 'ISC\Plugin::get_options' );
+			_deprecated_function( __METHOD__, '3.0.0', 'ISC\Options::get_options' );
 
 			$this->options = get_option( 'isc_options', $this->default_options() );
