+++ lib/composer/autoload_classmap.php
@@ -7,13 +7,18 @@ $baseDir = dirname($vendorDir);
 
 return array(
     'Composer\\InstalledVersions' => $vendorDir . '/composer/InstalledVersions.php',
+    'ISC\\Admin' => $baseDir . '/admin/admin.php',
+    'ISC\\Admin_Utils' => $baseDir . '/admin/utils.php',
     'ISC\\Autoloader' => $baseDir . '/includes/class-autoloader.php',
     'ISC\\Feedback' => $baseDir . '/includes/feedback.php',
     'ISC\\Helpers' => $baseDir . '/includes/helpers.php',
+    'ISC\\Image_Sources\\Admin' => $baseDir . '/includes/image-sources/admin.php',
     'ISC\\Image_Sources\\Analyze_HTML' => $baseDir . '/includes/image-sources/analyze-html.php',
     'ISC\\Image_Sources\\Image_Sources' => $baseDir . '/includes/image-sources/image-sources.php',
+    'ISC\\Image_Sources\\Utils' => $baseDir . '/includes/image-sources/utils.php',
     'ISC\\Indexer' => $baseDir . '/includes/indexer.php',
     'ISC\\Newsletter' => $baseDir . '/includes/newsletter.php',
+    'ISC\\Options' => $baseDir . '/includes/options.php',
     'ISC\\Plugin' => $baseDir . '/includes/plugin.php',
     'ISC\\Pro\\Custom_Attribute_Processor' => $baseDir . '/pro/includes/Custom_Attribute_Processor.php',
     'ISC\\Pro\\IPTC' => $baseDir . '/pro/includes/IPTC.php',
@@ -30,7 +35,6 @@ return array(
     'ISC\\Standard_Source' => $baseDir . '/includes/standard-source.php',
     'ISC\\Unused_Images' => $baseDir . '/includes/unused-images.php',
     'ISC\\User' => $baseDir . '/includes/user.php',
-    'ISC_Admin' => $baseDir . '/admin/admin.php',
     'ISC_Block_Options' => $baseDir . '/includes/block-options/block-options.php',
     'ISC_Class' => $baseDir . '/includes/isc.class.php',
     'ISC_Log' => $baseDir . '/includes/log.php',
