+++ lib/composer/autoload_static.php
@@ -27,13 +27,18 @@ class ComposerStaticInitdb72a6bac11cb0e6b971811c93d09a9b
 
     public static $classMap = array (
         'Composer\\InstalledVersions' => __DIR__ . '/..' . '/composer/InstalledVersions.php',
+        'ISC\\Admin' => __DIR__ . '/../..' . '/admin/admin.php',
+        'ISC\\Admin_Utils' => __DIR__ . '/../..' . '/admin/utils.php',
         'ISC\\Autoloader' => __DIR__ . '/../..' . '/includes/class-autoloader.php',
         'ISC\\Feedback' => __DIR__ . '/../..' . '/includes/feedback.php',
         'ISC\\Helpers' => __DIR__ . '/../..' . '/includes/helpers.php',
+        'ISC\\Image_Sources\\Admin' => __DIR__ . '/../..' . '/includes/image-sources/admin.php',
         'ISC\\Image_Sources\\Analyze_HTML' => __DIR__ . '/../..' . '/includes/image-sources/analyze-html.php',
         'ISC\\Image_Sources\\Image_Sources' => __DIR__ . '/../..' . '/includes/image-sources/image-sources.php',
+        'ISC\\Image_Sources\\Utils' => __DIR__ . '/../..' . '/includes/image-sources/utils.php',
         'ISC\\Indexer' => __DIR__ . '/../..' . '/includes/indexer.php',
         'ISC\\Newsletter' => __DIR__ . '/../..' . '/includes/newsletter.php',
+        'ISC\\Options' => __DIR__ . '/../..' . '/includes/options.php',
         'ISC\\Plugin' => __DIR__ . '/../..' . '/includes/plugin.php',
         'ISC\\Pro\\Custom_Attribute_Processor' => __DIR__ . '/../..' . '/pro/includes/Custom_Attribute_Processor.php',
         'ISC\\Pro\\IPTC' => __DIR__ . '/../..' . '/pro/includes/IPTC.php',
@@ -50,7 +55,6 @@ class ComposerStaticInitdb72a6bac11cb0e6b971811c93d09a9b
         'ISC\\Standard_Source' => __DIR__ . '/../..' . '/includes/standard-source.php',
         'ISC\\Unused_Images' => __DIR__ . '/../..' . '/includes/unused-images.php',
         'ISC\\User' => __DIR__ . '/../..' . '/includes/user.php',
-        'ISC_Admin' => __DIR__ . '/../..' . '/admin/admin.php',
         'ISC_Block_Options' => __DIR__ . '/../..' . '/includes/block-options/block-options.php',
         'ISC_Class' => __DIR__ . '/../..' . '/includes/isc.class.php',
         'ISC_Log' => __DIR__ . '/../..' . '/includes/log.php',
