+++ tests/wpunit/Pblc/Captions/Add_Source_Captions_To_Content_Test.php
@@ -4,7 +4,7 @@ namespace ISC\Tests\WPUnit\Pblc\Captions;
 
 use ISC\Tests\WPUnit\WPTestCase;
 use ISC_Public;
-use ISC\Plugin;
+use ISC\Options;
 
 /**
  * Test ISC_Public::add_source_captions_to_content()
@@ -133,7 +133,7 @@ class Add_Source_Captions_To_Content_Test extends WPTestCase {
 	 */
 	public function test_image_without_pretext() {
 		// empty the pretext
-		$isc_options                   = Plugin::get_options();
+		$isc_options                   = Options::get_options();
 		$isc_options['source_pretext'] = '';
 		update_option( 'isc_options', $isc_options );
 
@@ -148,7 +148,7 @@ class Add_Source_Captions_To_Content_Test extends WPTestCase {
 	 */
 	public function test_image_with_caption_style_none() {
 		// set the caption style to none
-		$isc_options                  = Plugin::default_options();
+		$isc_options                  = Options::default_options();
 		$isc_options['caption_style'] = 'none';
 		update_option( 'isc_options', $isc_options );
